Efi hello

+++ source/os_specific/efi/oseficlib.c
@@ -125,6 +125,17 @@
 
 #define ACPI_EFI_PRINT_LENGTH   256
 
+#define ACPI_EFI_KEY_ESC        0x0000
+#define ACPI_EFI_KEY_BACKSPACE  0x0008
+#define ACPI_EFI_KEY_ENTER      0x000D
+#define ACPI_EFI_KEY_CTRL_C     0x0003
+
+#define ACPI_EFI_ASCII_NULL     0x00
+#define ACPI_EFI_ASCII_DEL      0x7F
+#define ACPI_EFI_ASCII_ESC      0x1B
+#define ACPI_EFI_ASCII_CR       '\r'
+#define ACPI_EFI_ASCII_NL       '\n'
+
 
 /* Local prototypes */
 
@@ -142,6 +153,11 @@ AcpiEfiConvertArgcv (
     char                    ***ArgvPtr,
     char                    **BufferPtr);
 
+static int
+AcpiEfiGetFileInfo (
+    FILE                    *File,
+    ACPI_EFI_FILE_INFO      **InfoPtr);
+
 static CHAR16 *
 AcpiEfiFlushFile (
     FILE                    *File,
@@ -153,6 +169,7 @@ AcpiEfiFlushFile (
 
 /* Local variables */
 
+FILE                        *stdin = NULL;
 FILE                        *stdout = NULL;
 FILE                        *stderr = NULL;
 static ACPI_EFI_FILE_HANDLE AcpiGbl_EfiCurrentVolume = NULL;
@@ -160,6 +177,7 @@ ACPI_EFI_GUID               AcpiGbl_LoadedImageProtocol = ACPI_EFI_LOADED_IMAGE_
 ACPI_EFI_GUID               AcpiGbl_TextInProtocol = ACPI_SIMPLE_TEXT_INPUT_PROTOCOL;
 ACPI_EFI_GUID               AcpiGbl_TextOutProtocol = ACPI_SIMPLE_TEXT_OUTPUT_PROTOCOL;
 ACPI_EFI_GUID               AcpiGbl_FileSystemProtocol = ACPI_SIMPLE_FILE_SYSTEM_PROTOCOL;
+ACPI_EFI_GUID               AcpiGbl_GenericFileInfo = ACPI_EFI_FILE_INFO_ID;
 
 int                         errno = 0;
 
@@ -324,7 +342,8 @@ fclose (
     ACPI_EFI_FILE_HANDLE    EfiFile;
 
 
-    if (File == stdout || File == stderr)
+    if (File == stdin || File == stdout ||
+        File == stderr)
     {
         return;
     }
@@ -332,6 +351,135 @@ fclose (
     (void) uefi_call_wrapper (AcpiGbl_EfiCurrentVolume->Close, 1, EfiFile);
 
     return;
+}
+
+
+/*******************************************************************************
+ *
+ * FUNCTION:    fgetc
+ *
+ * PARAMETERS:  File                - File descriptor
+ *
+ * RETURN:      The character read or EOF on the end of the file or error
+ *
+ * DESCRIPTION: Read a character from the file.
+ *
+ ******************************************************************************/
+
+int
+fgetc (
+    FILE                    *File)
+{
+    UINT8                   Byte;
+    int                     Length;
+
+
+    Length = fread (ACPI_CAST_PTR (void, &Byte), 1, 1, File);
+    if (Length == 0)
+    {
+        Length = EOF;
+    }
+    else if (Length == 1)
+    {
+        Length = (int) Byte;
+    }
+
+    return (Length);
+}
+
+
+/*******************************************************************************
+ *
+ * FUNCTION:    fputc
+ *
+ * PARAMETERS:  File                - File descriptor
+ *              c                   - Character byte
+ *
+ * RETURN:      The character written or EOF on the end of the file or error
+ *
+ * DESCRIPTION: Write a character to the file.
+ *
+ ******************************************************************************/
+
+int
+fputc (
+    FILE                    *File,
+    char                    c)
+{
+    UINT8                   Byte = (UINT8) c;
+    int                     Length;
+
+
+    Length = fwrite (ACPI_CAST_PTR (void, &Byte), 1, 1, File);
+    if (Length == 0)
+    {
+        Length = EOF;
+    }
+    else if (Length == 1)
+    {
+        Length = (int) Byte;
+    }
+
+    return (Length);
+}
+
+
+/*******************************************************************************
+ *
+ * FUNCTION:    fgets
+ *
+ * PARAMETERS:  File                - File descriptor
+ *
+ * RETURN:      The string read
+ *
+ * DESCRIPTION: Read a string from the file.
+ *
+ ******************************************************************************/
+
+char *
+fgets (
+    char                    *s,
+    ACPI_SIZE               Size,
+    FILE                    *File)
+{
+    ACPI_SIZE               ReadBytes = 0;
+    int                     Ret;
+
+
+    if (Size <= 1)
+    {
+        errno = EINVAL;
+        return (NULL);
+    }
+    while (ReadBytes < (Size - 1))
+    {
+        Ret = fgetc (File);
+        if (Ret == EOF)
+        {
+            if (ReadBytes == 0)
+            {
+                return (NULL);
+            }
+            break;
+        }
+        else if (Ret < 0)
+        {
+            errno = EIO;
+            return (NULL);
+        }
+        else if (Ret == '\n')
+        {
+            s[ReadBytes++] = (char) Ret;
+            break;
+        }
+        else
+        {
+            s[ReadBytes++] = (char) Ret;
+        }
+    }
+
+    s[ReadBytes] = '\0';
+    return (s);
 }
 
 
@@ -357,14 +505,87 @@ fread (
     ACPI_SIZE               Count,
     FILE                    *File)
 {
-    int                     Length = -1;
-    ACPI_EFI_FILE_HANDLE    EfiFile;
-    UINTN                   ReadSize;
-    ACPI_EFI_STATUS         EfiStatus;
+    int                         Length = -EINVAL;
+    ACPI_EFI_FILE_HANDLE        EfiFile;
+    ACPI_SIMPLE_INPUT_INTERFACE *In;
+    UINTN                       ReadSize;
+    ACPI_EFI_STATUS             EfiStatus;
+    ACPI_EFI_INPUT_KEY          Key;
+    ACPI_SIZE                   Pos = 0;
 
+
+    if (!Buffer)
+    {
+        errno = EINVAL;
+        goto ErrorExit;
+    }
+
+    ReadSize = Size * Count;
 
     if (File == stdout || File == stderr)
     {
+        /* Do not support read operations on output console */
+    }
+    else if (File == stdin)
+    {
+        In = ACPI_CAST_PTR (ACPI_SIMPLE_INPUT_INTERFACE, File);
+
+        while (Pos < ReadSize)
+        {
+WaitKey:
+            EfiStatus = uefi_call_wrapper (In->ReadKeyStroke, 2, In, &Key);
+            if (ACPI_EFI_ERROR (EfiStatus))
+            {
+                if (EfiStatus == ACPI_EFI_NOT_READY)
+                {
+                    goto WaitKey;
+                }
+                errno = EIO;
+                Length = -EIO;
+                fprintf (stderr,
+                    "SIMPLE_INPUT_INTERFACE->ReadKeyStroke() failure.\n");
+                goto ErrorExit;
+            }
+
+            switch (Key.UnicodeChar)
+            {
+            case ACPI_EFI_KEY_CTRL_C:
+
+                break;
+
+            case ACPI_EFI_KEY_ENTER:
+
+                *(ACPI_ADD_PTR (UINT8, Buffer, Pos)) = (UINT8) ACPI_EFI_ASCII_CR;
+                if (Pos < ReadSize - 1)
+                {
+                    /* Drop CR in case we don't have sufficient buffer */
+
+                    Pos++;
+                }
+                *(ACPI_ADD_PTR (UINT8, Buffer, Pos)) = (UINT8) ACPI_EFI_ASCII_NL;
+                Pos++;
+                break;
+
+            case ACPI_EFI_KEY_BACKSPACE:
+
+                *(ACPI_ADD_PTR (UINT8, Buffer, Pos)) = (UINT8) ACPI_EFI_ASCII_DEL;
+                Pos++;
+                break;
+
+            case ACPI_EFI_KEY_ESC:
+
+                *(ACPI_ADD_PTR (UINT8, Buffer, Pos)) = (UINT8) ACPI_EFI_ASCII_ESC;
+                Pos++;
+                break;
+
+            default:
+
+                *(ACPI_ADD_PTR (UINT8, Buffer, Pos)) = (UINT8) Key.UnicodeChar;
+                Pos++;
+                break;
+            }
+        }
+        Length = Pos;
     }
     else
     {
@@ -374,7 +595,6 @@ fread (
             errno = EINVAL;
             goto ErrorExit;
         }
-        ReadSize = Size * Count;
 
         EfiStatus = uefi_call_wrapper (AcpiGbl_EfiCurrentVolume->Read, 3,
             EfiFile, &ReadSize, Buffer);
@@ -382,6 +602,7 @@ fread (
         {
             fprintf (stderr, "EFI_FILE_HANDLE->Read() failure.\n");
             errno = EIO;
+            Length = -EIO;
             goto ErrorExit;
         }
         Length = ReadSize;
@@ -417,12 +638,19 @@ AcpiEfiFlushFile (
     CHAR16                  *Pos,
     BOOLEAN                 FlushAll)
 {
+    ACPI_SIMPLE_TEXT_OUTPUT_INTERFACE    *Out;
 
-    if (FlushAll || Pos >= (End - 1))
+
+    if (File == stdout || File == stderr)
     {
-        *Pos = 0;
-        uefi_call_wrapper (File->OutputString, 2, File, Begin);
-        Pos = Begin;
+        Out = ACPI_CAST_PTR (ACPI_SIMPLE_TEXT_OUTPUT_INTERFACE, File);
+
+        if (FlushAll || Pos >= (End - 1))
+        {
+            *Pos = 0;
+            uefi_call_wrapper (Out->OutputString, 2, Out, Begin);
+            Pos = Begin;
+        }
     }
 
     return (Pos);
@@ -451,7 +679,7 @@ fwrite (
     ACPI_SIZE               Count,
     FILE                    *File)
 {
-    int                     Length = -1;
+    int                     Length = -EINVAL;
     CHAR16                  String[ACPI_EFI_PRINT_LENGTH];
     const char              *Ascii;
     CHAR16                  *End;
@@ -462,7 +690,11 @@ fwrite (
     ACPI_EFI_STATUS         EfiStatus;
 
 
-    if (File == stdout || File == stderr)
+    if (File == stdin)
+    {
+        /* Do not support write operations on input console */
+    }
+    else if (File == stdout || File == stderr)
     {
         Pos = String;
         End = String + ACPI_EFI_PRINT_LENGTH - 1;
@@ -516,11 +748,68 @@ ErrorExit:
 
 /*******************************************************************************
  *
+ * FUNCTION:    AcpiEfiGetFileInfo
+ *
+ * PARAMETERS:  File                - File descriptor
+ *              InfoPtr             - Pointer to contain file information
+ *
+ * RETURN:      Clibrary error code
+ *
+ * DESCRIPTION: Get file information.
+ *
+ ******************************************************************************/
+
+static int
+AcpiEfiGetFileInfo (
+    FILE                    *File,
+    ACPI_EFI_FILE_INFO      **InfoPtr)
+{
+    ACPI_EFI_STATUS         EfiStatus = ACPI_EFI_BUFFER_TOO_SMALL;
+    ACPI_EFI_FILE_INFO      *Buffer = NULL;
+    UINTN                   BufferSize = SIZE_OF_ACPI_EFI_FILE_INFO + 200;
+    ACPI_EFI_FILE_HANDLE    EfiFile;
+
+
+    if (!InfoPtr)
+    {
+        errno = EINVAL;
+        return (-EINVAL);
+    }
+
+    while (EfiStatus == ACPI_EFI_BUFFER_TOO_SMALL)
+    {
+        EfiFile = (ACPI_EFI_FILE_HANDLE) File;
+        Buffer = AcpiOsAllocate (BufferSize);
+        if (!Buffer)
+        {
+            errno = ENOMEM;
+            return (-ENOMEM);
+        }
+        EfiStatus = uefi_call_wrapper (EfiFile->GetInfo, 4, EfiFile,
+            &AcpiGbl_GenericFileInfo, &BufferSize, Buffer);
+        if (ACPI_EFI_ERROR (EfiStatus))
+        {
+            AcpiOsFree (Buffer);
+            if (EfiStatus != ACPI_EFI_BUFFER_TOO_SMALL)
+            {
+                errno = EIO;
+                return (-EIO);
+            }
+        }
+    }
+
+    *InfoPtr = Buffer;
+    return (0);
+}
+
+
+/*******************************************************************************
+ *
  * FUNCTION:    ftell
  *
  * PARAMETERS:  File                - File descriptor
  *
- * RETURN:      Size of current position
+ * RETURN:      current position
  *
  * DESCRIPTION: Get current file offset.
  *
@@ -531,8 +820,32 @@ ftell (
     FILE                    *File)
 {
     long                    Offset = -1;
+    UINT64                  Current;
+    ACPI_EFI_STATUS         EfiStatus;
+    ACPI_EFI_FILE_HANDLE    EfiFile;
 
 
+    if (File == stdin || File == stdout || File == stderr)
+    {
+        Offset = 0;
+    }
+    else
+    {
+        EfiFile = (ACPI_EFI_FILE_HANDLE) File;
+
+        EfiStatus = uefi_call_wrapper (EfiFile->GetPosition, 2,
+            EfiFile, &Current);
+        if (ACPI_EFI_ERROR (EfiStatus))
+        {
+            goto ErrorExit;
+        }
+        else
+        {
+            Offset = (long) Current;
+        }
+    }
+
+ErrorExit:
     return (Offset);
 }
 
@@ -557,8 +870,59 @@ fseek (
     long                    Offset,
     int                     From)
 {
+    ACPI_EFI_FILE_INFO      *Info;
+    int                     Error;
+    ACPI_SIZE               Size;
+    UINT64                  Current;
+    ACPI_EFI_STATUS         EfiStatus;
+    ACPI_EFI_FILE_HANDLE    EfiFile;
 
-    return (-1);
+
+    if (File == stdin || File == stdout || File == stderr)
+    {
+        return (0);
+    }
+    else
+    {
+        EfiFile = (ACPI_EFI_FILE_HANDLE) File;
+        Error = AcpiEfiGetFileInfo (File, &Info);
+        if (Error)
+        {
+            return (Error);
+        }
+        Size = Info->FileSize;
+        AcpiOsFree (Info);
+
+        if (From == SEEK_CUR)
+        {
+            EfiStatus = uefi_call_wrapper (EfiFile->GetPosition, 2,
+                EfiFile, &Current);
+            if (ACPI_EFI_ERROR (EfiStatus))
+            {
+                errno = ERANGE;
+                return (-ERANGE);
+            }
+            Current += Offset;
+        }
+        else if (From == SEEK_END)
+        {
+            Current = Size - Offset;
+        }
+        else
+        {
+            Current = Offset;
+        }
+
+        EfiStatus = uefi_call_wrapper (EfiFile->SetPosition, 2,
+            EfiFile, Current);
+        if (ACPI_EFI_ERROR (EfiStatus))
+        {
+            errno = ERANGE;
+            return (-ERANGE);
+        }
+    }
+
+    return (0);
 }
 
 
@@ -829,8 +1193,13 @@ efi_main (
 
     ST = SystemTab;
     BS = SystemTab->BootServices;
-    stdout = SystemTab->ConOut;
-    stderr = SystemTab->ConOut;
+    stdin = ACPI_CAST_PTR (ACPI_EFI_FILE, SystemTab->ConIn);
+    stdout = ACPI_CAST_PTR (ACPI_EFI_FILE, SystemTab->ConOut);
+    stderr = ACPI_CAST_PTR (ACPI_EFI_FILE, SystemTab->ConOut);
+
+    /* Disable the platform watchdog timer if we go interactive */
+
+    uefi_call_wrapper(BS->SetWatchdogTimer, 4, 0, 0x0, 0, NULL);
 
     /* Retrieve image information */
